Add MVC tests to ExperimentsConditionsDetailsController class
