+++ app/src/main/java/uk/ac/ebi/atlas/ebeyedump/ExperimentsConditionsDetailsController.java
@@ -1,6 +1,7 @@
 package uk.ac.ebi.atlas.ebeyedump;
 
 import org.springframework.context.annotation.Scope;
+import org.springframework.http.HttpHeaders;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import uk.ac.ebi.atlas.model.experiment.Experiment;
@@ -37,6 +38,7 @@ public class ExperimentsConditionsDetailsController {
     public void generateTsvFormatBaseline(HttpServletResponse response) {
         writeTsvLinesToResponse(
                 response,
+                "assaygroupsdetails.tsv",
                 experiment -> new BaselineExperimentAssayGroupsLines(
                         (BaselineExperiment) experiment, getExperimentDesign(experiment.getAccession())),
                 RNASEQ_MRNA_BASELINE,
@@ -48,6 +50,7 @@ public class ExperimentsConditionsDetailsController {
     public void generateTsvFormatDifferential(HttpServletResponse response) {
         writeTsvLinesToResponse(
                 response,
+                "contrastdetails.tsv",
                 experiment -> new DifferentialExperimentContrastLines(
                         (DifferentialExperiment) experiment, getExperimentDesign(experiment.getAccession())),
                 MICROARRAY_1COLOUR_MICRORNA_DIFFERENTIAL,
@@ -61,12 +64,15 @@ public class ExperimentsConditionsDetailsController {
     }
 
     private void writeTsvLinesToResponse(HttpServletResponse response,
-                                         Function<Experiment, Iterable<String[]>> linesIteratorProducer,
+                                         String fileName,
+                                         Function<Experiment<?>, Iterable<String[]>> linesIteratorProducer,
                                          ExperimentType... experimentTypes) {
+        response.setStatus(HttpServletResponse.SC_OK);
+        response.addHeader(HttpHeaders.CONTENT_DISPOSITION, "attachment;filename=" + fileName);
         response.setContentType("text/tab-separated-values");
 
         try {
-            for (Experiment experiment : experimentTrader.getPublicExperiments(experimentTypes)) {
+            for (Experiment<?> experiment : experimentTrader.getPublicExperiments(experimentTypes)) {
                 for (String[] line : linesIteratorProducer.apply(experiment)) {
                     response.getWriter().write(join("\t", line) + "\n");
                 }
