--- app/src/test/java/uk/ac/ebi/atlas/ebeyedump/ExperimentsConditionsDetailsControllerTest.java
@@ -1,45 +1,0 @@
-package uk.ac.ebi.atlas.ebeyedump;
-
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.context.web.WebAppConfiguration;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import uk.ac.ebi.atlas.configuration.TestConfig;
-
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@ExtendWith(SpringExtension.class)
-@WebAppConfiguration
-@ContextConfiguration(classes = TestConfig.class)
-class ExperimentsConditionsDetailsControllerTest {
-
-    @Autowired
-    private WebApplicationContext wac;
-
-    private MockMvc mockMvc;
-
-    @BeforeEach
-    void setUp() {
-        this.mockMvc = MockMvcBuilders.webAppContextSetup(this.wac).build();
-    }
-
-
-    @Test
-    void generateTsvFormatBaseline() throws Exception {
-        this.mockMvc.perform(get("/api/assaygroupsdetails.tsv"))
-            .andExpect(status().isOk());
-    }
-
-    @Test
-    void generateTsvFormatDifferential() throws Exception {
-        this.mockMvc.perform(get("/api/contrastdetails.tsv"))
-            .andExpect(status().isOk());
-    }
-}
